requirements.txt
langgraph 
langchain 
langchain-community 
langchain-core 
langchain-ollama 
duckduckgo-search
ollama
langchain-mcp-adapters
mcp
wikipedia-api
websockets